Simple formatting adjustments
