+++ .editorconfig
@@ -3,10 +3,10 @@ root = true
 
 [*] 
 indent_style = space 
-file_header_template = Copyright (c) IEvangelist. All rights reserved.\nLicensed under the MIT License.
+file_header_template = Copyright © IEvangelist. All rights reserved.\nLicensed under the MIT License.
 
 [*.cs] 
-charset = utf-8
+charset = utf-8-bom
 end_of_line = crlf
 #indent_size = 4
 indent_style = tab
@@ -72,7 +72,7 @@ dotnet_naming_symbols.const_fields.applicable_accessibilities = *
 dotnet_naming_symbols.const_fields.applicable_kinds = field
 dotnet_naming_symbols.const_fields.required_modifiers = const
 dotnet_separate_import_directive_groups = true
-dotnet_sort_system_directives_first = false
+dotnet_sort_system_directives_first = true
 dotnet_style_coalesce_expression = true:suggestion
 dotnet_style_collection_initializer = true:suggestion
 dotnet_style_explicit_tuple_names = true:suggestion
