+++ Microsoft.Azure.CosmosRepository/src/DefaultRepositoryFactory.cs
@@ -1,30 +1,31 @@
-﻿// Copyright (c) IEvangelist. All rights reserved.
+﻿// Copyright © IEvangelist. All rights reserved.
 // Licensed under the MIT License.
 
 using System;
+
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.Azure.CosmosRepository
 {
-    /// <inheritdoc/>
-    internal class DefaultRepositoryFactory : IRepositoryFactory
-    {
-        readonly IServiceProvider _serviceProvider;
+	/// <inheritdoc/>
+	internal class DefaultRepositoryFactory : IRepositoryFactory
+	{
+		readonly IServiceProvider _serviceProvider;
 
-        /// <summary>
-        /// Constructor for the default respositroy factory.
-        /// </summary>
-        /// <param name="serviceProvider"></param>
-        public DefaultRepositoryFactory(IServiceProvider serviceProvider) =>
-            _serviceProvider = serviceProvider
-                ?? throw new ArgumentNullException(
-                    nameof(serviceProvider),
-                    "A service provider instance is required.");
+		/// <summary>
+		/// Constructor for the default respositroy factory.
+		/// </summary>
+		/// <param name="serviceProvider"></param>
+		public DefaultRepositoryFactory(IServiceProvider serviceProvider) =>
+			_serviceProvider = serviceProvider
+				?? throw new ArgumentNullException(
+					nameof(serviceProvider),
+					"A service provider instance is required.");
 
-        /// <inheritdoc/>
-        public IRepository<TItem> RepositoryOf<TItem>()
-            where TItem : Item =>
-            _serviceProvider.GetRequiredService<IRepository<TItem>>();
+		/// <inheritdoc/>
+		public IRepository<TItem> RepositoryOf<TItem>()
+			where TItem : Item =>
+			_serviceProvider.GetRequiredService<IRepository<TItem>>();
 
-    }
+	}
 }
