+++ Microsoft.Azure.CosmosRepository/src/Extensions/ExpressionExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) IEvangelist. All rights reserved.
+﻿// Copyright © IEvangelist. All rights reserved.
 // Licensed under the MIT License.
 
 using System;
@@ -8,62 +8,38 @@ using System.Linq.Expressions;
 
 namespace Microsoft.Azure.CosmosRepository.Extensions
 {
-    /// <summary>
-    /// Borrowed from:
-    /// https://docs.microsoft.com/en-us/archive/blogs/meek/linq-to-entities-combining-predicates
-    /// </summary>
-    internal static class ExpressionExtensions
-    {
-        internal static Expression<T> Compose<T>(
-            this Expression<T> first,
-            Expression<T> second,
-            Func<Expression, Expression, Expression> merge)
-        {
-            Dictionary<ParameterExpression, ParameterExpression> map =
-                first.Parameters
-                    .Select((parameter, index) => (parameter, second: second.Parameters[index]))
-                    .ToDictionary(p => p.second, p => p.parameter);
+	/// <summary>
+	/// Borrowed from:
+	/// https://docs.microsoft.com/en-us/archive/blogs/meek/linq-to-entities-combining-predicates
+	/// </summary>
+	internal static class ExpressionExtensions
+	{
+		internal static Expression<T> Compose<T>(
+			this Expression<T> first,
+			Expression<T> second,
+			Func<Expression, Expression, Expression> merge)
+		{
+			Dictionary<ParameterExpression, ParameterExpression> map =
+				first.Parameters
+					.Select((parameter, index) => (parameter, second: second.Parameters[index]))
+					.ToDictionary(p => p.second, p => p.parameter);
 
-            Expression secondBody = ParameterRebinder.ReplaceParameters(map, second.Body);
+			Expression secondBody = ParameterRebinder.ReplaceParameters(map, second.Body);
 
-            return Expression.Lambda<T>(merge(first.Body, secondBody), first.Parameters);
+			return Expression.Lambda<T>(merge(first.Body, secondBody), first.Parameters);
 
-        }
+		}
 
-        internal static Expression<Func<T, bool>> AndAlso<T>(
-            this Expression<Func<T, bool>> first,
-            Expression<Func<T, bool>> second) => first.Compose(second, Expression.AndAlso);
+		internal static Expression<Func<T, bool>> AndAlso<T>(
+			this Expression<Func<T, bool>> first,
+			Expression<Func<T, bool>> second) => first.Compose(second, Expression.AndAlso);
 
-        internal static Expression<Func<T, bool>> And<T>(
-            this Expression<Func<T, bool>> first,
-            Expression<Func<T, bool>> second) => first.Compose(second, Expression.And);
+		internal static Expression<Func<T, bool>> And<T>(
+			this Expression<Func<T, bool>> first,
+			Expression<Func<T, bool>> second) => first.Compose(second, Expression.And);
 
-        internal static Expression<Func<T, bool>> Or<T>(
-            this Expression<Func<T, bool>> first,
-            Expression<Func<T, bool>> second) => first.Compose(second, Expression.Or);
-    }
-
-    internal class ParameterRebinder : ExpressionVisitor
-    {
-        readonly Dictionary<ParameterExpression, ParameterExpression> _map;
-
-        internal ParameterRebinder(Dictionary<ParameterExpression, ParameterExpression> map) =>
-            _map = map ?? new Dictionary<ParameterExpression, ParameterExpression>();
-
-        internal static Expression ReplaceParameters(
-            Dictionary<ParameterExpression, ParameterExpression> map, Expression exp) =>
-            new ParameterRebinder(map).Visit(exp);
-
-        /// <inheritdoc />
-        protected override Expression VisitParameter(ParameterExpression parameter)
-        {
-            if (_map.TryGetValue(parameter, out ParameterExpression replacement))
-            {
-                parameter = replacement;
-            }
-
-            return base.VisitParameter(parameter);
-        }
-
-    }
+		internal static Expression<Func<T, bool>> Or<T>(
+			this Expression<Func<T, bool>> first,
+			Expression<Func<T, bool>> second) => first.Compose(second, Expression.Or);
+	}
 }
